Add files via upload
Team Name: Fairview
Students Name: Md.Saiful Islam
Students Id:2118CSE50227
Project Title: Scientific calculator (SCLTR)
Category: Desktop Software
GitHub Repository: Saiful18038/Scientific-calculator (github.com)

Project Description
Scientific calculator (SCLTR) is a python project for teacher's or student's for presenting their
project without any hassle. With this software student only need to submit their presentation in
PDF format, the SCLTR will do the rest like presenting documentation and report in PDF
format, downloading code from GitHub and present them by compiling it automatically. If any
libraries are required SCLTR will download them allative

The scientific calculator is a type of an electronic calculator in which different calculating
methods are involved. The methods contain scientific, mathematical and some methods related to
engineering. These types of calculators are used when one is going for a higher education,
because higher education contains these types of essential elements. Other than that, the
scientific calculators have more features as compared to the ordinary calculator in which many
kinds of calculations can be performed. The functions which are involved in this type of
calculator are scientific notation, floating point values, logarithmic functions, exponential
functions, complex numbers, fractions, statistics, programmability, equation solving, matrix
calculations, calculus, conversion of units and physical constants. The scientific calculator is
used in various fields for example in astronomy, geology, physics, chemistry, somehow in
biology as well. The calculators are mostly used by the students studying in schools, colleges,
universities. This calculator will not only be performing the basic functions but also the complex
calculations.
Calculator is a python project. For developing this project, I have used some framework and
libraries other than the core python. For Graphical User Interface (GUI) I have used " Tkinter " ,
a Python's standard GUI framework. It is popular for its simplicity and graphical user interface;
it is open source and available under the Python License. One of the advantages of choosing
Tkinter is that since it comes by default, there is an abundance of resources, both codes and
reference books. Also, with the community being old and active.

For PDF viewer and scraping data from the PDF file I have used a python library called "
PyPDF2 " . It is a pure python library so it can run on any platform without any platform -
related dependencies on any external libraries.
Project Outcome
The project scientific calculator will give various outputs. It will be a bug free calculator. You
will be able to understand the purpose of calculator. You will be easily understanding the basic

functions on the provided calculator. You will be able to understand which type of problem do
you have and then you will be easily solving the given calculations along with the method which
is given to you. Not only this, you will also be able to know whether you are performing right or
wrong.

The Project Presentation Software (SCLTR) will be available in both Windows and Linux
Platforms .

Features: When a PDF file or a link to the PDF file is added to the SCLTR it will read the PDF
and will get all necessary resources needed for that project. There will be no need for additional
software like PDF reader, Python IDE for presenting the project. All the necessary tools are
included in the Scientific calculator